+++ docs/CHANGELOG.rst
@@ -33,6 +33,7 @@ Changed
   ``resolwebio/rnaseq:6.0.0`` to match those from
   ``resolwebio/common:3.0.1`` Docker image and update genome-tools to
   1.6.2 in ``resolwebio/rnaseq:6.0.0`` Docker image
+- Add Java memory settings to processes ``bqsr`` and ``markduplicates``
 
 Fixed
 -----